projeto_alterado/requirements.txt
asgiref==3.8.1
chardet==5.2.0
crispy-bootstrap5==2024.2
Django==5.0.6
django-cors-headers==4.4.0
django-crispy-forms==2.2
djangorestframework==3.15.2
et-xmlfile==1.1.0
openpyxl==3.1.5
pillow==10.4.0
reportlab==4.2.2
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
